requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
azure-ai-vision==0.11.1b1
azure-cognitiveservices-speech==1.29.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.28
azure-core==1.28.0
azure-identity==1.13.0
azure-mgmt-core==1.4.0
azure-mgmt-web==7.1.0
azure-storage-blob==12.17.0
beautifulsoup4==4.12.2
black==23.3.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
cfgv==3.3.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.1.0
coverage==7.2.3
cryptography==41.0.1
cssselect==1.2.0
cycler==0.11.0
dataclasses-json==0.5.7
decorator==5.1.1
distlib==0.3.6
dnspython==2.3.0
docker==6.0.1
duckduckgo-search==2.8.6
exceptiongroup==1.1.1
filelock==3.11.0
flake8==6.0.0
Flask==2.3.2
fonttools==4.41.0
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-api-python-client==2.85.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
greenlet==2.0.2
gTTS==2.3.1
h11==0.14.0
httplib2==0.22.0
identify==2.5.22
idna==3.4
importlib-metadata==6.6.0
iniconfig==2.0.0
isodate==0.6.1
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
jwt==1.3.1
kiwisolver==1.4.4
langchain==0.0.189
loguru==0.7.0
lxml==4.9.2
markdown-it-py==2.2.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.2
mccabe==0.7.0
mdurl==0.1.2
msal==1.22.0
msal-extensions==1.0.0
msrest==0.7.1
multidict==6.0.4
mypy-extensions==1.0.0
nodeenv==1.7.0
numexpr==2.8.4
numpy==1.24.1
oauthlib==3.2.2
openai==0.27.7
openapi-schema-pydantic==1.2.4
opencensus==0.11.2
opencensus-context==0.1.3
opencensus-ext-azure==1.1.9
opencv-python==4.8.0.74
orjson==3.8.10
outcome==1.2.0
packaging==23.1
pandas==1.5.3
pathspec==0.11.1
pdfminer.six==20221105
pdfplumber==0.10.0
Pillow==9.5.0
pinecone-client==2.2.1
platformdirs==3.2.0
playsound==1.2.2
pluggy==1.0.0
portalocker==2.7.0
pre-commit==3.2.2
protobuf==4.23.2
psutil==5.9.5
pyarrow==12.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.8
pydeck==0.8.1b0
pydub==0.25.1
pyflakes==3.0.1
Pygments==2.15.1
PyJWT==2.7.0
Pympler==1.0.1
pyodbc==4.0.39
pyparsing==3.0.9
PyPDF2==3.0.1
pypdfium2==4.18.0
pyrsistent==0.19.3
PySocks==1.7.1
pytest==7.3.1
pytest-mock==3.10.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-pptx==0.6.21
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
readability-lxml==0.8.1
redis==4.5.4
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.4.1
rsa==4.9
seaborn==0.12.2
selenium==4.8.3
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sourcery==1.2.0
SQLAlchemy==2.0.15
streamlit==1.23.1
streamlit-chat==0.0.2.2
tabulate==0.9.0
tenacity==8.2.2
tiktoken==0.4.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
tweepy==4.13.0
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.3
uritemplate==4.1.1
urllib3>=1.26.0,<2.0.3
validators==0.20.0
virtualenv==20.21.0
watchdog==3.0.0
webdriver-manager==3.8.6
websocket-client==1.5.1
Werkzeug==2.3.6
win32-setctime==1.1.0
wsproto==1.2.0
XlsxWriter==3.1.2
yarl==1.9.2
zipp==3.15.0
